+++ src/main/java/com/VooTreeVeeVuu/domain/repository/FacilityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface FacilityRepository extends JpaRepository<Facility, Integer> {
+public interface FacilityRepository extends JpaRepository<Facility, Long> {
 //    List<Facility> findAll ();
 //
 //    Optional<Facility> findById (Integer id);
